# backend/resources/backend-openapi.yaml
openapi: "3.0.3"
info:
  title: "backend API"
  description: "backend API"
  version: "1.0.0"
servers:
  - url: "https://backend"
paths:
...
  /api/person/{id}:
    get:
      summary: "GET api/person/{id}"
      operationId: "getById"
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "number"
            format: "int64"
      responses:
        "200":
          description: "OK"
...
  /api/showtime/of-date-grouped/{dateStr}:
    get:
      summary: "GET api/showtime/of-date-grouped/{dateStr}"
      operationId: "getCurrentGrouped"
      parameters:
        - name: "dateStr"
          in: "path"
          required: true
          schema:
            type: "string"
      responses:
        "200":
          description: "OK"
  /api/showtime/of-date/{dateStr}:
    get:
      summary: "GET api/showtime/of-date/{dateStr}"
      operationId: "getOfDate"
      parameters:
        - name: "dateStr"
          in: "path"
          required: true
          schema:
            type: "string"
      responses:
        "200":
          description: "OK"
  /api/showtime/of-movie/{movieId}:
    get:
      summary: "GET api/showtime/of-movie/{movieId}"
      operationId: "getOfMovie"
      parameters:
        - name: "movieId"
          in: "path"
          required: true
          schema:
            type: "number"
            format: "int64"
      responses:
        "200":
          description: "OK"
...